Moving the class files to the class/ folder
