+++ packages/CCFG/src/class/CCFG.ts
