+++ packages/CCFG/src/class/Instruction.ts
