+++ packages/CCFG/src/class/Node.ts
