+++ packages/CCFG/src/index.ts
@@ -1,4 +1,4 @@
 // src/index.ts
-export * from './CCFG.js';
-export * from './Instruction.js';
-export * from './Node.js';
+export * from './class/CCFG.js';
+export * from './class/Instruction.js';
+export * from './class/Node.js';
