+++ packages/backend-compiler/src/class/TraversalContext.ts
